+++ src/index.js
@@ -2,10 +2,10 @@ import React from "react"
 import ReactDOM from "react-dom"
 import registerServiceWorker from "./registerServiceWorker"
 
-import Assemble from "./Assemble"
+import Assembly from "./Assembly"
 
 ReactDOM.render(
-  <Assemble link={"linked_hash"} />,
+  <Assembly link={"linked_hash"} />,
   document.getElementById('root'),
 )
